+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 
 * Support for Realm database storage - using `RealmStorage` class.
 
+## Changed
+
+* UIReactions now properly unwrap data models, even for cases when model contains double optional value.
+
 ## [4.4.2](https://github.com/DenHeadless/DTCollectionViewManager/releases/tag/4.4.2)
 
 ## Fixed